# server/openapi.yaml
openapi: 3.0.3
info:
  title: Code Interview Platform API
  description: |
    Backend API for the collaborative coding interview platform.
    
    This API supports:
    - Creating and managing interview sessions
    - Real-time code collaboration via WebSocket
    - Code execution in multiple programming languages
    - Session state persistence
    - User presence and awareness
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@codeinterviewplatform.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.codeinterviewplatform.com
    description: Production server
...
paths:
...
  /sessions/{sessionId}/participants:
    get:
      tags:
        - Collaboration
      summary: Get active participants
      description: Get list of currently active participants in a session
      operationId: getParticipants
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Unique session identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Participants list retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                  participants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Participant'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sessions/{sessionId}/history:
    get:
      tags:
        - Sessions
      summary: Get session history
      description: Retrieve edit history and snapshots for a session
      operationId: getSessionHistory
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Unique session identifier
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Maximum number of history entries to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Session history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/HistoryEntry'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Participant:
      type: object
      properties:
        userId:
          type: string
          description: Unique user identifier
          example: "user-abc123"
        name:
          type: string
          description: User display name
          example: "Anonymous User"
        joinedAt:
          type: string
          format: date-time
          description: When the user joined the session
        cursorPosition:
          type: object
          properties:
            line:
              type: integer
            column:
              type: integer
          description: Current cursor position in the editor
    HistoryEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: When this change was made
        userId:
          type: string
          description: User who made the change
        changeType:
          type: string
          enum: [edit, language-change, snapshot]
          description: Type of change
        description:
          type: string
          description: Human-readable description of the change
        codeSnapshot:
          type: string
          description: Code content at this point (for snapshots only)
          nullable: true
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: "SESSION_NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
          example: "The requested session does not exist"
        statusCode:
          type: integer
          description: HTTP status code
          example: 404
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional JWT authentication for protected sessions

# Optional security for future implementation
# security:
#   - bearerAuth: []